add conditional for okta feature in config files
