+++ apps/boilerplate-front-end/src/App.jsx
@@ -1,4 +1,4 @@
-import { useEffect } from 'react'
+import { useEffect, useState } from 'react'
 
 //Recoil states & values
 import { useRecoilValue } from 'recoil'
@@ -16,12 +16,37 @@ import PredictUserProfileProvider from './components/predict/PredictUserProfileP
 // Import Components
 import Main from './Main'
 
+// Okta Import for authentication
+import config from '@/config/oktaLogin'
+import { OktaAuth, toRelativeUrl } from '@okta/okta-auth-js'
+import { Security } from '@okta/okta-react'
+
+import { useNavigate } from 'react-router-dom'
+import { shouldHaveOktaLogin } from './config/featuresConfig'
+
 const App = () => {
+  const [oktaAuth, setOktaAuth] = useState(null)
+
   // Index to make the main search queries
   const index = useRecoilValue(mainIndex)
 
   // Get userID from Predict
   const userId = useRecoilValue(predictUserIdAtom)
+
+  const shouldHaveLoginWithOkta = useRecoilValue(shouldHaveOktaLogin)
+
+  // Create a callback authentication with OKTA
+  const navigate = useNavigate()
+  const restoreOriginalUri = (_oktaAuth, originalUri) => {
+    navigate(toRelativeUrl(originalUri || '/', window.location.origin))
+  }
+
+  useEffect(() => {
+    if (shouldHaveLoginWithOkta) {
+      const oktaAuth = new OktaAuth(config.oidc)
+      setOktaAuth(oktaAuth)
+    }
+  }, [])
 
   useEffect(() => {
     window.process = {
@@ -35,7 +60,13 @@ const App = () => {
         indexName={index}
         routing={true}
       >
-        <Main />
+        {shouldHaveLoginWithOkta ? (
+          <Security oktaAuth={oktaAuth} restoreOriginalUri={restoreOriginalUri}>
+            <Main />
+          </Security>
+        ) : (
+          <Main />
+        )}
       </InstantSearch>
     </PredictUserProfileProvider>
   )
