+++ apps/boilerplate-front-end/src/config/featuresConfig.js
@@ -129,3 +129,8 @@ export const shouldHaveCartFunctionality = atom({
   key: 'shouldHaveCartFunctionality', // unique ID (with respect to other atoms/selectors)
   default: true, // default value (aka initial value)
 })
+
+export const shouldHaveOktaLogin = atom({
+  key: 'shouldHaveOktaLogin', // unique ID (with respect to other atoms/selectors)
+  default: false, // default value (aka initial value)
+})
